Lots of bug fixes, add emoticons widget support

+++ mod/src/main/java/tv/twitch/android/mod/models/UserInfoCallback.java
@@ -1,5 +1,5 @@
 package tv.twitch.android.mod.models;
 
 public interface UserInfoCallback {
-    void userInfo(String userName, long userId);
+    void userInfo(String userName, int userId);
 }
